Checked in by Sameer.
There disease tab (under create study) was failing as the meddra_versions were by default being read and passed back to the client. In the case when there is no meddra_version imported, this was causing an indexOutOfBound exception. This change populates the same only if the study is configured to have meddra terminology.

SVN-Revision: 4881