BUG - 139
Checked in by Sameer.
Modified the groovy files related to Meddra import to support backward compatibility. The idea is to start the sequences created on the ids of the meddra concept tables from a value higher than the meddra_code for the respective table. This way the meddra already loaded and new versions can co-exist.

SVN-Revision: 4885