CAAERS-305
Checked in by Sameer.
Initially the fieldgroups that was holding the meddra_versions values was declared to be a class member in the details_tab. Only if it was null, the values were repopulated from DB. As a result of this the new/deleted versions were not being shown on the page. To fix this, the fieldgroup is now defined as a local variable and it will always be populated with the fresh values from DB before rendering the page.

SVN-Revision: 5186