CAAERS-1622
In RoutingAndReview page when a person with role AeCoordinator tried viewing the aeReport in pdf format using the dropdown (under format column), he used to get an access denied error. But as a part of new feature, the ae-coordinator can click on the aeReport and view the report in pdf(readonly) format. So this column was itself removed from the page. Hence the fix.

SVN-Revision: 8526