CAAERS-2490
There was a circular dependency when trying to access adverseEventRoutingAndReviewRepository from reportRepository. So had to split ReportRepository into logical groups. Created a new service class called ReportValidationService which supports the methods like "isSubmittable" to test the completeness of the reports while the methods to create/ amend/ withdraw reports are retained in the ReportRepository service class.

SVN-Revision: 11197