CAAERS-3856
As per this fix - If the logged in person is SAECoordinator at the coordinating center
and the workflow has ended ( => The user has no action to take), still it will be retained
in the results list of routing and review search query. Also now if the Report review
searches for the "Review Status = Ready to submit .." he will get results. Earlier no
results were returned as there was no action available for the reviewer.

SVN-Revision: 12479